refactor: 포인트 클래스 패키지변경
