--- src/main/java/kr/hhplus/be/server/application/point/PointFacade.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.application.point;
-
-public class PointFacade {
-}
