+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/PointResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.api.point.response;
+package kr.hhplus.be.server.interfaces.api.point;
 
 public record PointResponse(long id, long amount, long userId) {
 }
