+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/PointUseRequest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.api.point.request;
+package kr.hhplus.be.server.interfaces.api.point;
 
 
 public record PointUseRequest(Long userId, Long amount) {
